Improve directories structure by moving data parsing sources to the same one level of abstraction
