+++ scripts/fifaDataParser/constants.js
