+++ scripts/fifaDataParser/parseWithArray.js
