+++ scripts/fifaDataParser/parseWithRegExp.js
