+++ scripts/fifaDataParser/utilities.js
