+++ scripts/index.js
@@ -1,7 +1,7 @@
 const axios = require("axios").default;
 const path = require("path");
-const { parseText } = require("./parseWithArray");
-const { writeDataToJSONFiles } = require("./utilities");
+const { parseText } = require("./fifaDataParser/parseWithArray");
+const { writeDataToJSONFiles } = require("./fifaDataParser/utilities");
 
 const URL =
   "https://www.ea.com/games/fifa/fifa-22/news/fifa-22-all-leagues-clubs-teams-list?setLocale=en-us";
